modularized API data queries
