+++ Makefile.am
@@ -293,6 +293,10 @@ API_PLUGIN_FILES = \
     web/api/exporters/allmetrics.h \
     web/api/exporters/shell/allmetrics_shell.c \
     web/api/exporters/shell/allmetrics_shell.h \
+    web/api/queries/rrdr.c \
+    web/api/queries/rrdr.h \
+    web/api/queries/query.c \
+    web/api/queries/query.h \
     $(NULL)
 
 STREAMING_PLUGIN_FILES = \
