+++ configure.ac
@@ -607,6 +607,7 @@ AC_CONFIG_FILES([
     web/api/exporters/Makefile
     web/api/exporters/shell/Makefile
     web/api/exporters/prometheus/Makefile
+    web/api/queries/Makefile
     web/gui/Makefile
     web/server/Makefile
     web/server/single/Makefile
