+++ web/api/rrd2json.h
@@ -5,6 +5,7 @@
 
 #include "web_api_v1.h"
 #include "exporters/allmetrics.h"
+#include "queries/rrdr.h"
 
 #define HOSTNAME_MAX 1024
 
@@ -42,23 +43,6 @@
 #define GROUP_MAX               3
 #define GROUP_SUM               4
 #define GROUP_INCREMENTAL_SUM   5
-
-#define RRDR_OPTION_NONZERO         0x00000001 // don't output dimensions will just zero values
-#define RRDR_OPTION_REVERSED        0x00000002 // output the rows in reverse order (oldest to newest)
-#define RRDR_OPTION_ABSOLUTE        0x00000004 // values positive, for DATASOURCE_SSV before summing
-#define RRDR_OPTION_MIN2MAX         0x00000008 // when adding dimensions, use max - min, instead of sum
-#define RRDR_OPTION_SECONDS         0x00000010 // output seconds, instead of dates
-#define RRDR_OPTION_MILLISECONDS    0x00000020 // output milliseconds, instead of dates
-#define RRDR_OPTION_NULL2ZERO       0x00000040 // do not show nulls, convert them to zeros
-#define RRDR_OPTION_OBJECTSROWS     0x00000080 // each row of values should be an object, not an array
-#define RRDR_OPTION_GOOGLE_JSON     0x00000100 // comply with google JSON/JSONP specs
-#define RRDR_OPTION_JSON_WRAP       0x00000200 // wrap the response in a JSON header with info about the result
-#define RRDR_OPTION_LABEL_QUOTES    0x00000400 // in CSV output, wrap header labels in double quotes
-#define RRDR_OPTION_PERCENTAGE      0x00000800 // give values as percentage of total
-#define RRDR_OPTION_NOT_ALIGNED     0x00001000 // do not align charts for persistant timeframes
-#define RRDR_OPTION_DISPLAY_ABS     0x00002000 // for badges, display the absolute value, but calculate colors with sign
-#define RRDR_OPTION_MATCH_IDS       0x00004000 // when filtering dimensions, match only IDs
-#define RRDR_OPTION_MATCH_NAMES     0x00008000 // when filtering dimensions, match only names
 
 extern void rrd_stats_api_v1_chart(RRDSET *st, BUFFER *wb);
 extern void rrd_stats_api_v1_charts(RRDHOST *host, BUFFER *wb);
